APICLIENT-3: Change version ranges as suggested by Michael Markl My reasoning for it is: Once a dependent (of this lib) needs "entities", but in version "1.1.1", because there was some misbehavior introduced in "1.1.2", then this would collide with our lib if we set entities to "^1.1.2".
According to semver, new minor versions are backwards compatible while
adding new functionalities and new patch versions are only used for
bug-fixes (and new major versions for breaking changes of course).

I changed my mind to the following: If we don't want to test our lib to
work with all minor versions <1.2.x, we shouldn't set the version range
to "1.x.x" (because 1.2.0 could have introduced some new functionality
that we depend on), but what we can do is set it to "^1.2" instead.
So we can specify that our lib depends on the API of major version 1 and
minor version 2 (which allows any minor version >= 2 according to
semver) --- while ignoring the actual patch version: that's up to the
dependent of our lib whether he uses the latest patches of "entities"
or not